Break integration test peer endpoint
